build: produce static libraries
This also adds the constructor to the plugin class, as it should now be
called to register the qml types.

Dynamic linking mostly has its problem on Sailfish OS, as it creates
larger binaries and RPMs. Since this is the only application consuming
it, I see no benefit in using it.